+++ src/components/common/Card.scss
